perf(encoder): batch VLC + UVLC emissions in HT cleanup encode
Combine the 3 separate emitVLCBits() calls per quad pair (quad 0 VLC,
quad 1 VLC, UVLC) into a single call with a pre-combined 64-bit
codeword. Widen emitVLCBits() parameter to uint64_t to accommodate
the combined value (max ~34 bits).

Reduces VLC emit+drain cycles by 2/3 per quad pair.  For a 4K image
this eliminates ~16K drain-check iterations.

Measured on ElephantDream_4K (8-bit) and StainedGlass_4K (12-bit),
single-thread, median of 15 warm runs vs main baseline:

  ED lossless  209.8 → 200.4 ms  (-4.5%)
  SG lossless  191.8 → 185.1 ms  (-3.5%)
  ED lossy     102.2 → 100.7 ms  (-1.4%)

AVX2 variant only; other SIMD variants unchanged.

+++ source/core/coding/ht_block_encoding_avx2.hpp
@@ -187,9 +187,6 @@ class state_MS_enc {
         emit_dword();
       }
     }
-    // while (ctreg >= 64) {
-    //   emit_qword();
-    // }
 #endif
   }
 
@@ -250,8 +247,8 @@ class state_VLC_enc {
     buf[pos + 1] = 0xFF;
   }
 
-  FORCE_INLINE void emitVLCBits(uint32_t cwd, uint32_t len) {
-    Creg |= static_cast<uint64_t>(cwd) << ctreg;
+  FORCE_INLINE void emitVLCBits(uint64_t cwd, uint32_t len) {
+    Creg |= cwd << ctreg;
     ctreg += len;
     while (ctreg >= 32) {
       emit_dword();
